Some refactoring (clarify the folder structure)
