+++ bin/models/ean_product.dart
