+++ bin/models/receipt_product.dart
