+++ bin/specific/s_kaupat/save_rps_as_csv.dart
@@ -1,6 +1,6 @@
 import 'dart:io';
-import 'helper.dart';
-import 'receipt_product.dart';
+import '../../utils/helper.dart';
+import '../../models/receipt_product.dart';
 
 /// Saves receipt products as a CSV file.
 void receiptProducts2CSV(List<ReceiptProduct> products, String csvFilePath) {
@@ -13,8 +13,6 @@ void receiptProducts2CSV(List<ReceiptProduct> products, String csvFilePath) {
         '${product.name};${product.quantity};${product.pricePerUnit};${product.totalPrice};${product.discountCounted}\n');
   }
 
-  var date = formattedDateTime();
-
-  var file = File('$csvFilePath/receipt_products_$date.csv');
+  var file = File('$csvFilePath/receipt_products_${formattedDateTime()}.csv');
   file.writeAsString(csv.toString());
 }
