+++ bin/utils/shop_selector_helper.dart
